LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-align-text-align-expected.txt

foo bar abcdefghijklmnopqrstuvwxyz
foo bar abcdefghijklmnopqrstuvwxyz
foo bar abcdefghijklmnopqrstuvwxyz
foo bar abcdefghijklmnopqrstuvwxyz
foo bar abcdefghijklmnopqrstuvwxyz

PASS legend
FAIL legend[align=left] assert_equals: expected "start" but got "left"
FAIL legend[align=center] assert_equals: expected "start" but got "center"
FAIL legend[align=right] assert_equals: expected "start" but got "right"
FAIL legend[align=justify] assert_equals: expected "start" but got "justify"
